feat: waku/config
